Add files via upload
A salesman is given a list of locations in the Traveling Salesman issue (TSP), a 
classic optimization issue. His task is to determine the shortest route that visits 
each city exactly once and returns to the beginning city. The optimal solution of 
TSP for a large number of cities can be computationally costly. In this example, I'll 
give you a Prolog code that uses a brute-force method to solve a simple TSP 
instance. Remember that large instances of the problem are inefficient using this 
code.