Add files via upload
The traversal problem for a robot is one of the many challenges that can be solved 
using the Means-End Analysis (MEA), a popular approach in artificial intelligence. 
In this case, the robot must use all available tools or actions to go from its current 
position to a target place.
Assume you have a straightforward grid environment with movable robot that can 
move left, right, up, and down, and that its objective is to go to a predetermined 
spot.